feat: enhance consistent hash load balancing with weights
Update consistent hash load balancing to support weighted backends.
The `weight` parameter is now utilized in both `round_robin` and
`consistent_hash`
algorithms, ensuring requests are distributed proportionally to backend
weights.

This change also includes documentation updates for the `weight`
directive
and internal code adjustments to facilitate weighted distribution in the
consistent hash ring. A new test case has been added to verify the
weighted
distribution behavior of the consistent hash ring.

+++ modules/http-proxy/src/upstream/lb/round_robin.rs
@@ -75,7 +75,7 @@ impl WeightedRoundRobinState {
             #[cfg(not(target_has_atomic = "64"))]
             let current_weight = old_weight + *weight as i32;
             #[cfg(target_has_atomic = "64")]
-            let current_weight = old_weight as i64 + *weight as i64;
+            let current_weight = old_weight + *weight as i64;
             if current_weight > best_weight {
                 best_weight = current_weight;
                 best_index = i;
